requirements.txt
--find-links https://download.pytorch.org/whl/cpu/torch
torch==2.6.0+cpu

anthropic==0.49.0
cohere==5.19.0
datasets==4.2.0
fastapi==0.115.12
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
groq==0.32.0
httpx==0.28.1
huggingface-hub==0.35.3
infisicalsdk==1.0.11

langchain==0.3.25
langchain-astradb==0.6.1
langchain-cohere==0.4.6
langchain-experimental==0.3.4
langchain-google-genai==2.1.12
langchain-groq==0.3.8
langchain-huggingface
langchain-milvus==0.2.1
langchain-mistralai==0.2.12
langchain-postgres==0.0.16
langchain-text-splitters==0.3.8
langgraph-sdk==0.2.9
langsmith
logfire==4.14.1
loguru==0.7.3

opentelemetry-api==1.38.0
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-alephalpha==0.39.1
opentelemetry-instrumentation-anthropic==0.39.1
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-bedrock==0.39.1
opentelemetry-instrumentation-chromadb==0.39.1
opentelemetry-instrumentation-cohere==0.39.1
opentelemetry-instrumentation-crewai==0.39.1
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-google-generativeai==0.39.1
opentelemetry-instrumentation-groq==0.39.1
opentelemetry-instrumentation-haystack==0.39.1
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-instrumentation-lancedb==0.39.1
opentelemetry-instrumentation-langchain==0.39.1
opentelemetry-instrumentation-llamaindex==0.39.1
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-marqo==0.39.1
opentelemetry-instrumentation-milvus==0.39.1
opentelemetry-instrumentation-mistralai==0.39.1
opentelemetry-instrumentation-ollama==0.39.1
opentelemetry-instrumentation-openai==0.39.1
opentelemetry-instrumentation-pinecone==0.39.1
opentelemetry-instrumentation-qdrant==0.39.1
opentelemetry-instrumentation-replicate==0.39.1
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-sagemaker==0.39.1
opentelemetry-instrumentation-sqlalchemy==0.59b0
opentelemetry-instrumentation-system-metrics==0.59b0
opentelemetry-instrumentation-threading
opentelemetry-instrumentation-together==0.39.1
opentelemetry-instrumentation-transformers==0.39.1
opentelemetry-instrumentation-urllib3
opentelemetry-instrumentation-vertexai==0.39.1
opentelemetry-instrumentation-watsonx==0.39.1
opentelemetry-instrumentation-weaviate==0.39.1
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.59b0

nltk==3.9.1
numpy==2.3.4
pandas==2.3.3
pgvector==0.3.6
pillow==12.0.0
praw==7.8.1
prawcore==2.4.0
psycopg2-binary==2.9.11
psycopg-pool==3.2.6
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0
pyjwt==2.10.1
pymilvus==2.6.2
pymongo==4.15.3
pymupdf==1.24.13
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
safetensors==0.6.2
sentence-transformers==5.1.1
sqlalchemy==2.0.44
tiktoken==0.9.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
uvicorn==0.34.1